Disables banner
This removes the banner from the top of the page since we have already
had the webinar. Instead of removing the code, we leave the partial and
a comment with how to turn it back on for future instances where we
would want to display a message at the top of the page.

Signed-off-by: Christopher Webber <cwebber@chef.io>