Fail travis build on failure to deploy website
On Travis, if your `after_script` fails (exits non-zero), the build
still gets to be green.

This makes it so the building of the website and the deploying (if we're
on the right branch) are all in the `script` instead, so if the deploy
fails, the Travis build goes red.

(This happened: https://travis-ci.org/habitat-sh/habitat/jobs/141353230.
It was because we made a change on S3, and the script failed, but we
weren't notified by a Travis failure, but by somebody noticing their
change had not been deployed.)

Signed-off-by: Nathan L Smith <smith@chef.io>