Set the SSL_CERT_FILE for the director
We need to set the SSL_CERT_FILE in the environment in order for the
processes started by the director to know where to find it, so that
HTTPS requests can have the SSL certificates verified. This change was
done manually on "live." When we set up the "acceptance" environment,
the Sign In process was throwing a 500 that looks like this:

ERROR:habitat_builder_api::http::handlers: github authentication, err=HyperError(Ssl(OpenSslErrors([UnknownError { library: "SSLroutines", function: "ssl3_get_server_certificate", reason: "certificate verify failed" }])))

This isn't because we are using a self-signed certifcate on acceptance
though, it's because the application couldn't verify GitHub's
certificate because it didn't know about the CA certs.

This does require that the binlink is setup for `/bin/hab`, which is
done by our terraform provisioning.

Signed-off-by: jtimberman <joshua@chef.io>