[core] Set mtime on files during unpacking archive
By default libarchive ignores the time data in the archive when
unpacking files. This means that the mtime of the files on disk after
install are all set to the time of install. This causes problems with
packages like guile which look at the mtime of its compiled assets to
determine if it needs to recompile them at startup.

Signed-off-by: Steven Danna <steve@chef.io>