[hab,hab-sup] Fix progress bar/newline rendering issue.
This change addresses an issue where a download/upload progress bar
doesn't print a newline character once complete and thus appends the
next line to itself. This is observed when resizing terminal windows,
copying command line output, and occasionally when a command is
executing in a non-interactive shell.

Our progress bar implementation uses the `pbr` crate and gets combined
into a broadcast writer or tee reader before calling `io::copy` on an
IO stream. Rather than include extra external calls once the copy is
completed, the progress is also kept inside the `ProgressBar` struct.
The last call to `write()` will cause a newline to be printed to the
`stdout` stream and then the stream will be flushed before proceeding.

It may be possible to add further smarts to the upstream library, but
this fix resolves the rendering issues in the meantime. Additional
changes to the command line's UI behavior are coming shortly which may
further tweak this implementation.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>