[hab] Add an optional arg to install.sh to specify a version of hab.
This is an optional/opt-in change to the `install.sh` script which helps
consumers specify a version (i.e. `install.sh 0.8.0`) or a release (i.e.
`install.sh 0.8.0/20160729175817`) to be installed. Note that the
version installed will originate from the Depot and not the Bintray
repo--the `hab` provided by the Bintray repo is only used to install the
target version of `hab`. The more you know...

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>