Quote no_proxy correctly
Docker for Mac 1.12.0-a build 11213 and above have started automatically
setting `no_proxy` for the local docker networks.

```bash
no_proxy="*.local, 169.254/16, 192.168.99/24"
```

In `hab-studio`, we use that variable when we build up the list of
environment variables to pass to `env` before we exec the chroot. The
issue is quoting - the whitespace in the value causes the call to env
to see the parts as a new set of environment variables. While you would
think you could work around that by just making some judicious use of
`"` or `'` everywhere, you would be wrong - bash somehow knows, deep in
its dark soul, that we're being naughty.

The final answer was to do what all good shell scripts do. Throw up your
hands in anger and just use sed.

```bash
env="$env no_proxy=$(echo $no_proxy | $bb sed 's/, /,/g')"
```

Because it turns out the whitespace is totally un-needed in the
`no_proxy` value.

![I Fixed It!](http://gif.co/xlVR.gif)

Signed-off-by: Adam Jacob <adam@chef.io>