[http-client] Add missing colon on Host header when connecting to proxy.
When a tunneling proxy connection is established, a client makes a
request to the proxy with something like the following:

```
CONNECT willem.habitat.sh:443 HTTP/1.1
Host: willem.habitat.sh:443
Proxy-Authorization: Basic amRvZTpzYXNzYWZyYXNz
```

The HTTP verb in this case is `CONNECT` and what follows are HTTP
headers, namely the `Host` and optionally the `Proxy-Authorization`
headers.

This change adds a missing colon to the `Host` header which causes some
proxy servers to correctly fail with an invalid request (i.e. an
HTTP/400).

Special thanks to the heroic efforts of @hiucimon who lent his effort,
time, and was able to diagnose this on a proper network setup.

Closes #1180

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>