Quick fix for sorting package identifiers returned by the Depot API
This is a first pass implementation to get a fix out the door. We're
pulling a bulk of strings out of the database and allocating identifiers
for each then performing a sort in memory. This is going to be alright
for now, but in the future we should use weights in Redis' sorted sets
and adjust the rank of all identifiers on insertion of a new package
so we only perform the sorting overhead on write instead of each read

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>