Move buffer trimming to the final buffer
We used to have a bug, and that bug was that we weren't reading an
entire file from the tarball; isntead, we were only getting the very
first chunk.

We fixed that bug. Yay us!

However, we wanted all the data we read from our metadata files to have
no trailing whitespace or newlines (because hey - you really just want
the data you want, not the extra garbage on the end that humans like.
Silly humans.) We did this trimming on the content of the block we read.

When you then generate a file that has the buffer boundary pefectly
aligned with a newline, you will get your newline chomped off, and an
error from habitat that your metadata is malformed.

This commit fixes that issue by moving the trim to the final buffer,
rather than on every call.

Signed-off-by: Adam Jacob <adam@chef.io>