Iterate on plans with --config-from
This commit enables the `--config-from` option to the supervisor. It
allows you to pass the directory where your habitat plan lives, and when
you run the supervisor, it will use the configuration it finds there
rather than the configuration in the package.

While this violates a bunch of useful properties in production, it's
very useful for plan creation. Here is an example:

```bash
$ build redis
[.. hab builds stuff ..]
$ hab start core/redis
[ .. redis starts ..]
```

At this point, you might realize a hook is wrong, or a configuration
value isn't set. Currently, testing this would either require you to
edit the installed path in `/hab/pkgs`, or recompile all of redis. With
this option, you can instead say:

```bash
$ hab start core/redis --config-from /src
```

And edit the configuration file in place, test to your hearts content,
and then issue one final build when you're sure eveything works.

In future work, it would be cool to include some file watching logic, to
automatically reload the service when your configuration changes. For
now, though, this works great, and it doesn't take any new dependencies.

Signed-off-by: Adam Jacob <adam@chef.io>