[hab,sup] Add common UI subsystem to manage output display.
This change is a first move to add a consistent "design language" to the
user-facing tools--primarily the `hab` program. It introduces a `UI`
struct in the common crate which knows how to write headings, status
messages, etc in the context of consoles that could be interactive or
noninteractive and terminals that support color or ones that do not
support color.

Most of the work to date is focused around proper terminal capability
detection (via the `term` crate) and interactive status (via the
`libc::isatty` and `kernel32::GetConsoleMode` calls). More testing and
validation against Windows platforms is required, but the minimum work
is there.

The `UI` struct has several methods on it that help authors of `hab`
subcommands present a consistent response to an end user. The `.begin()`
and `.end()` methods properly decorate and color the respective
messages, and the `.status()` method helps steer the subcommand author
into a verb-first message with consistent symbol decoration.

A pseudo-code example might be:

```
try!(ui.begin(format!("Installing {}", &pkg_ident)));
// install package and determine fully qualified identifier
// ...
let fqpi = ...;
try!(ui.status(Status::Downloading, &fqpi));
// ...
try!(ui.begin(format!("Install of {} successful", &fqpi)));
```

which might yield output similar to:

```
» Installing core/jq-static
↓ Downloading core/jq-static/1.5/20160730001211
...
★ Install of core/jq-static/1.5/20160730001211 sucessful
```

The `UI` struct is responsible for determining coloring capability,
context, and strategy, which symbol decorations to use by default, etc.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>