[hab/mac] Prompt for core origin key when performing a Mac hab release.
This change alters the ordering of the `mac-build.sh` script to prompt
the user for a core origin signing key as soon as possible in the
process. As a consequence, a `wget` command is no longer present for the
mac `install.sh` program and therefore `curl` is used instead (Homebrew
installs the `wget` command whereas `curl` is available on stock Mac
installations).

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>