Worker building plans from GitHub source
This first pass will allow a worker to setup a workspace per job at
/hab/svc/hab-builder-worker/data/${jobid}. Inside this workspace
directory are three additional files/directories:

  * `${jobdir}/src` - where the source to be built for the plan is
    cloned to
  * `${jobdir}/studio` - the chroot where the build will occur
  * `${jobdir}/build.log` - the output from the build

If a build succeeds it will automatically attempt to upload to the
public depot. If the upload fails, the job will fail.

This PR does not include build log streaming, build log archival,
or any of the automatic dependency rebuild or automatic notification
of build from GitHub features. Those will be coming later.

We also assume that you are building core plans only. A strategy for
retrieving secret keys and building other origin's plans is to come.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>