Set sensible defaults for pkg_pconfig_dirs plan variable
Prior to these changes the default value for pkg_pconfig_dirs was an
empty array of strings. A plan author could override this variable
with directories containing `.pc` files which output the metafile
`PKG_CONFIG_PATH`. This path is propagated throughout the build
process making linking software easier and less error prone in Habitat.
This is useful, but we can do better.

We will now provide a better default if the plan author does not set
`pkg_pconfig_dirs` by attempting to locate a directory at
`$pkg_prefix/lib/pkgconfig` or `$pkg_prefix/share/pkgconfig` and, if
present, set the `pkg_pconfig_dirs` value to one or both of those
locations.

While `lib/pkgconfig` and `share/pkgconfig` are not documented as the
conventional locations by the pkgconfig documentation, it is referred to
as the conventional location by a number of hits from a google search.

This will make linking to software packaged by Habitat easier even if
the plan author didn't explicitly think of documenting pkgconfig dirs.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>