[core] Add core `become_command()` to abstract Unix/Windows exec diffs.
This change extracts the formerly named `exec_command()` function from
hab's `exec` module into a common interface `become_command()` in core's
newly created `os::process` module. The intent of this function
philisophically is to turn the current process into some other process
that takes over for the duration of the new process. In other words, the
existing process ceases to do any work.

There are 2 implementations for this function, which depend on the
underlying target platform:

Unix (i.e. Linux, Mac, BSD, etc)
--------------------------------

This implementation uses [`execvp(3)`][] which is now provided by the
[`exec()`][] function on Rust's stdlib [`CommandExt`][] trait. This
insures that there isn't a deeply nested process chain and that signals
such as `Ctrl+C` correctly kill all related processes. If a developer
needs to create a child process and supervise it or act on the child's
results then this is **not** the appropriate API to invoke, hence the
name `become_command()`.

[`execvp(3)`]: https://linux.die.net/man/3/execvp
[`exec()`]:
https://doc.rust-lang.org/std/os/unix/process/trait.CommandExt.html#tymethod.exec
[`CommandExt`]:
https://doc.rust-lang.org/std/os/unix/process/trait.CommandExt.html

Windows
-------

This implementation uses Rust's stdlib [`Command`] to execute the
command as a child process and then exits with the child's exit code.
Due to the lack of `exec*()` behavior. Note that there is an older
`_execv()` in a POSIX compatibility layer in MSVCRT, but the native
approach is to use a child process here, and so that's what we do.
Also note that while the `hab` process is not replaced by the new
process, we explicitly terminate the `hab` process after the child
process completes. If a developer needs to create a child process and
supervise it or act on the child's results then this is **not** the
appropriate API to invoke, hence the name `become_command()`.

[`Command`]:
https://doc.rust-lang.org/std/process/struct.Command.html#method.exec

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>