[hab] Attempt to call `sudo` on `hab studio *` for non-root users.
This change only currently affects Linux non-root users who attempt to
run the following subcommands:

* `hab studio *`
* `hab pkg build *`

The behavior introduced here will attempt to re-exec the command with
prepended with a `sudo -p "[sudo hab-studio] password for %u:" -E` in
front of the original `hab` subcommand. If the `sudo` command cannot be
found on the `PATH` then an error message is displayed to the user and
the program terminates with a non-zero exit code. The end result is
that, for non-root users on systems using sudoers:

    hab pkg build ...

Will be the same as running:

    sudo hab pkg build ...

Examples
--------

Running as a non-root user with appropriate permissions in sudoers:

```
> hab studio enter
[sudo hab-studio] password for jdoe:
   hab-studio: Creating Studio at /hab/studios/home--jdoe (default)
   hab-studio: Importing jdoe secret origin key
» Importing origin key from standard input
★ Imported secret origin key jdoe-20161005175939.
   hab-studio: Entering Studio at /hab/studios/home--jdoe (default)
   hab-studio: Exported: HAB_ORIGIN=jdoe

[1][default:/src:0]#
```

Running as a non-root user when the `sudo` command cannot be found on
`PATH`:

```
> sudo mv /usr/bin/sudo /usr/bin/ssudo

> hab studio enter
∅ Could not find the `sudo' command on the filesystem or in PATH.
∅ Running Habitat Studio requires root or administrator privileges. Please retry this command as a super user or use a privilege-granting facility such as sudo.

✗✗✗
✗✗✗ Root or administrator permissions required to complete operation
✗✗✗
```

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>