[ci] Remove unused and redundant cache directories.
This change removes the `$HOME/rust` directory which appears to be
long-unused (introduced in #513), and also removes the `$HOME/.cargo`
and `target` entries which appear to be
[handled](https://docs.travis-ci.com/user/caching/#Rust-Cargo-cache) by
the `cargo: true` lines in the .travis.yml.

This should reduce the time spent downloading (extra?) caches tarballs
off s3 and marginally speed up the build times.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>