[JEX-346] Unify the Linux and MacOS install scripts
We borrow some idioms and helpers from the omnitruck install.sh script
to make a better cross-platform install.sh script for Linux and MacOS.
This allows us to have a single installation script that handles both
platforms, and allows for easier extension to other platforms in the
future.

As the "mac" install.sh will no longer be necessary, it is removed
from the repository.

Signed-off-by: Joshua Timberman <joshua@chef.io>