[core] Fix hyper test compilation churn.
This fixes an issue that we see in our CI system and occasionally in
development. It turns out that the core component built the hyper crate
differently in test mode (as a dev-dependencies entry) than the rest of
the components in either test or normal mode.

Building the sequence of "LIB" crates (as provided in our Makefile)
leads eventually to a build of hyper that is different than the version
of hyper which was built alongside the iron crate (used in the depot and
vanilla http client crates). This behavior goes back almost to June, but
I posit that in combination with our Travis caching and the
short-circuiting logic present in `./support/ci/fast_pass.sh` has masked
this issue for months. Oi.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>