Make `fix_interpreter` resolve symbolic links
Currently if fix_interpreter runs into a symlink, it will fix the
interpreter, then replace the symlink with an actual copy of the fixed
file. This can lead to bad things if that's not what you wanted it to
do.

This will make it so if `fix_interpreter` encounters a symlink, it will
fix the file pointed to by the symlink instead of the symlink itself.

Signed-off-by: Nathan L Smith <smith@chef.io>