[plan-build] Ensure that `TERMINFO` is not present for all builds.
This change addresses an issues introduced when every Studio instance
started exporting the `TERMINFO` environment variable which was in scope
when the build program ran. Specifically, the build time behavior of the
`core/ncurses` package is different with and without this environment
variable. When set, no `$pkg_prefix/share/terminfo` directory is created
and when not set, the directory is created.

While this particular environment variable affects one currently known
package, there may be others current or future which could have
different build behaviors with this variable. In an attempt to preserve
the build system's "brutal minimalism", we've opted to explicitly unset
this environment variable.

Note that this is a warning to any future environment variables that are
added to a Studio instance--it may be hard to determine the impact such
a variable may have so best err on being conservative and explicitly
unsetting when it is desirable/necessary.

References #1435

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>