[devshell] Add a non-root user, jdoe, for root/sudo checking codepaths.
An arbitrary user such as this ends up being used over and over in
testing situations that having a known non-root user is useful for
Habitat developers. The user is `jdoe` with a simple password of `1234`
and who is also in the `sudo` group (thus being able to use `sudo`
commands).

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>