Add hab-butterfly
Adds a hab-butterfly CLI, which is called by `hab` if you issue a
command that needs to access the ring.

Adds the `Client` to butterfly, which allows you to send rumors directly
to a Butterfly server, without becoming a member first.

Currently this supports only the `hab config apply` command; a separate
PR will include the support for configuration files.

This leaves some warnings in place that will be cleaned up in the second
PR, as the code paths may well get resurrected.

Signed-off-by: Adam Jacob <adam@chef.io>