No longer render `me` for services not running
Previously, the supervisor would always render something for `all.me`,
for every service, regardless of whether this supervisor actually has a
running service or not.

This change makes it so that we now render a `me` value only if the
current supervisor has an instance of a service running in a given
service group.

Signed-off-by: Adam Jacob <adam@chef.io>