Fixes comment about hab workaround and hard-coded platforms
The comment wasn't as clear as it should be to describe the
problem with the hab macro.

More importantly we were having some issues with the creation
of an array that contained the supported shells remaining
around for the static lifetime. So the easy thing to do was
to hard code the shells in the string.

Signed-off-by: Franklin Webber <franklin@chef.io>