Initial prototype of a PowerShell-based Windows build system.
This initial work is designed to show what a PowerShell-based build
program and PowerShell-based Plans would look like. It is not completely
functional and has serveral issues, however it demonstrates the user
experience of such a fully developed system.

The approach taken here is a spirited *transliteration* of the original
Bash-based build program and Plan format which is currently in use. In
that sense, the same build phases (i.e. `prepare`, `build`, `install`,
etc.) are used, the same metadata variable names are used, etc. A small
example of a build Plan is provided for the `hab` component although due
to build and linking complexities, this plan does not fully compile the
Rust project. Importantly, the directory structure under `HAB_ROOT` and
the internal packaging format of the Habitat Artifact (.hart) has been
preserved with the intention of resuing most of the existing logic
present in the codebase. The build program currently only has 2 external
program dependencies (`hab` for artifact signing and future package
installs and `7z` for creating tarballs and compressing with xz
compression) which are checked for when the program boots (in a similar
fashion to the Bash-based build program).

The result is a very familiar build experience for previous Habitat
users. Further feedback will be required to validate that this approach
seems natural to native Windows developers and administrators who desire
commandline-driven/scriptable automation.

Currently there are several unimplemented behaviors and build phases
such as:

* No downloading and archive extraction logic
* No dependency calculation or locking logic
* Empty defaults for `build` and `install` phases

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>