[sup] Reimplement recursive toml config merging.
This change reimplements the logic which is responsible for merging down
runtime, environment, user, and default config for a service. The
initial implementation was a shallow merge only, meaning that sub tables
where not merged with each other--only top level keys. The next
implementation built on this code by supporting deep merging of toml
tables if present. There were a few issues with this implementation when
different value types were present for the same key. Additionally, the
Rust match-based logic was extended to support this which resulted in a
harder to understand function. Lastly, using recursion as part of the
implementation introduced a potential attack vector where a user could
craft a data structure deep enough to overflow the stack and cause the
supervisor to unwind and terminate.

This current implementation reduces the merging logic to the following:

* There are 2 toml tables: `me` (the current, config which will be
  mutated in place) and `other` (a reference to a config which override
  the current config).
* For each key in `other`, perform one of two things:
* If the values at `key` in `me` and `other` both contains toml tables,
  then call `toml_merge_recurse` on those value tables to recursively
  merge them down.
* Otherwise copy the value of `other` at `key` into the value of `key`
  in `me`. It makes no difference if the value types differ or are the
  same--the `other` table will always override anything in `me`.

The merge logic also contains a maximum recursive depth check in an
attempt to prevent burning through all stack space.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>