Fix unexpectedly long thread sleep in Supervisor's main loop
If a service takes a long time to initialize we may have exceeded our
desired one second maximum tick rate. If we have, we end up with a time
in the past which turns into a negative integer. Casting a negative
integer to a u64 wraps the integer causing the thread to sleep for a
very, very, long time.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>