[hab/install.sh] Allow installation of specific hab releases.
This change fixes the logic in hab's `install.sh` script to support
installation of fully qualified hab versions (i.e.
`0.12.1/20161102212401`).

Prior to this change, specifying a `verion/release` format such as
`sh install.sh -v 0.12.1/20161102212401` resulted in a 404 error which
was caused because in Bintray the *version* string is `version-release`
rather than `version/release`.

Note that on Linux, the same version of the desired `hab` binary is
downloaded even though the Bintray binary is thrown away at the end of
the installation process--it is only used as a trusted binary to install
the full `core/hab` Habitat package from the Depot. This may help in the
future to test any possible breaking changes in `hab install` and `hab
pkg binlink` commands.

Finally note that *only* `version/release` strings are
supported--specifying only the Habitat version string component (i.e.
`0.12.1`) will not work correctly and requires much more Bintray API
querying--it simply is not worth the effort.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>