[hab] Disable `config apply` & `file upload` on Mac and Windows systems.
As of the 0.14.0 release, the two subcommands `config apply` and `file
upload` are provided by an accompanying Habitat package
`core/hab-butterfly` which is installed on first use similar to the
behavior of the `sup` and `studio` subcommands on Linux. As the Depot
software currently doesn't support multiple target systems we cannot
simply download the required package and proceed. As such, until we can
provide a better experience for this on Mac and Windows, a message will
be displayed to users explaining that these subcommands are not
currently supported.

The following would be displayed:

```
> target/debug/hab config apply
∅ Running `config apply` on this operating system is not currently supported. Try running this command again on a 64-bit Linux operating system.

✗✗✗
✗✗✗ Subcommand `config apply' not supported on this operating system
✗✗✗
```

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>