Unify `make bldr-run` Makefile target to work on native Linux & Mac.
This change brings the 2 Procfiles into one version that should work
under most circumstances under Linux and Mac. The notable exception is
for systems that already have a `redis-server` process running, such as
a Homebrew-installed package on macOS or a system-installed package on
Ubuntu for example.

The 2 remaining Makefile targets are:

* `bldr-run` - which launches the Builder services necessary for
  developing against. By default on the Mac this will run in a devshell
  using Docker and will be run natively on Linux without Docker.
* `bldr-shell` - which launches a devshell (i.e. a Docker container)
  with forwarded ports but doesn't run any of the services

As the `BUILDING.md` instructions describe how to work natively on Mac
(i.e. not using Docker but targeting macOS directly), you can set
`export IN_DOCKER=false` for `make bldr-run` to compile and run on the
Mac.

The driving user feature for this change was to be able to run `make
bldr-run` on a Linux host without needing to resort to the same
bind mount/Docker overhead which slows down the development cycle.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>