[bldr] Push common args into their subcommands to prevent mis-invoking.
This change clears up a usability issue that I stumbled upon myself,
initially in `core/bldr-api`. I wanted to use a custom config and ran
the following:

    bldr-api --config /path/to/config.toml start

To my confusion, no matter what I did, my config was not getting picked
up. Upon re-reading the code and double checking against the clap API, I
learned that the `global` attribute on an argument will leave the arg
match at the level it was provided by the user (and not also copied down
to the subcommand). This means that both of these invocations result in
a started service:

    bldr-api --config /path/to/config.toml start
    bldr-api start --config /path/to/config.toml

The catch is that only the second one actually applies the config. Oi.

This fix provided here drops the `global` attribute on the `Arg` and
pushes them down into the `start` subcommand. Currently all the Builder
services only provide a `start` subcommand but in the future additional
subcommands may be added. At this point there is some `start` hardcoding
that needs to be generalized anyway and any copious arg duplication can
be dealt with then.

The new behavior is as follows:

```sh
> bldr-api start --config /path/to/config.toml
```

```sh
> bldr-api --config /path/to/config.toml start
error: Found argument '--config' which wasn't expected, or isn't valid in this context

USAGE:
    bldr-api [SUBCOMMAND]

For more information try --help
```

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>