Resolve peer address before assigning it's IP to member
The previous behavour of a peer binding would not resolve a dns entry before trying to unwrap a string. This caused a thread panic that would kill the supervisor.
The new behavour will try and resolve the ip's for dns entries and die if it is unable to do so.

Signed-off-by: Elliott Davis <edavis@chef.io>