Fix Hook output and remove "output" from HealthCheck and Hook errors
This PR accomplishes two goals:

1. stdout and stderr from running hooks will one again correctly be
   piped to the supervisor's output and it's originating service
   group will be included in it's preamble
2. The "output" member on the HealthCheck and Hook errors has been
   removed. See the description below for details

Originally we were storing the stdout of a failed hook and, by proxy, a
health check as a member of the error struct. We can't do this because
we don't know exactly how large the contents of the output would be and
this could cause the supervisor to crash. The output from hooks are
piped to the supervisor's console and can be retrieved from there.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>