Remove hab-director component
It's a large part of Habitat's vision for it's supervisor to run
multiple services but, at launch, we were unable to deliver this
functionality in time. The best version of the feature was on the
chopping block because it would have required considerable refactoring
to the supervisor's codebase but it was important to the team to
communicate to our new user base that this was an intended use case
for Habitat's runtime.

We decided to create the Director as a stop-gap measure which would be
replaced after we were able to refactor the supervisor to a place where
it could host multiple services. Because of how quickly the Director
was pieced together, like any project, it suffered in quality. We are
nearly ready to begin work on the final pieces of the supervisor to make
it multi-tenant so we're removing the Director from the codebase and
documentation. We feel it's important to do so before the supervisor
support is ready because while the Directory conveys our intentions, it
doesn't meet the minimum requirements to be a useful component in the
Habitat toolchain and causes more confusion and harm than good.

We look forward to delivering multi-tenant support in the supervisor
in the very near future!

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>