Migrate from rustc-serialize to serde in preparation for Rust 1.15
Serde is the preferred crate for serialization in Rust but we were
unable to change our serialization implementation because it requires
unstable/nightly Rust.

This commit removes our requirement on rustc-serialize and replaces
any calls to it with the serde equivalent. We're able to switch to serde
ahead of the release of Rust 1.15 by leveraging the serde_codegen crate
and some tricks in our rustc build process.

After Rust 1.15 is released, we can rollback the tricks by moving all
of the struct definitions from the `serde_types.in.rs` files placed in
the `src` root of many of our crates and disabling the codegen process
in our `build.rs` files.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>