Add Package Exports to build and runtime systems
This commit introduces a new concept for Habitat packages to make
a subset of service configuration data available to consumers through
the `--bind` runtime flag of a service.

A package which declares one or more exports is known as a "producer"
and the service using a `--bind` runtime flag to connect to a service
group running the producing package is known as the "consumer".

A package declares an export with the newly created `pkg_exports` plan
variable. This is an associative array which maps the services
configuration into a public key space. This public key space acts as
a promise, or contract, between the producer and consumer to mitigate
changes in configuration structure between updates.

The public configuration of a bind can be accessed within a newly
created `.cfg` accessor in a member of a bind from within any of our
rendered templates.

* Add `pkg_exports` to plan syntax
* Add EXPORTS metadata file to `.hart` files
* Update service rumor to include *exported* configuration data
* Update census entry to contain configuration data from a service
  rumor
* Update pkg plan generator template to include `pkg_exports` stub
* Additional refactors for readability, memory efficiency, and log
  usability

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>